Refactor(templates/JS): Update all files after run eslint and add husky
