+++ templates/javascript/.eslintrc.js
@@ -4,43 +4,43 @@ const path = require('path');
 const prettierOptions = JSON.parse(fs.readFileSync(path.resolve(__dirname, '.prettierrc'), 'utf8'));
 
 module.exports = {
-	env: {
-		jest: true,
-		browser: true,
-		node: true,
-		es6: true,
-	},
+  env: {
+    jest: true,
+    browser: true,
+    node: true,
+    es6: true,
+  },
 
-	extends: ['plugin:react/recommended', 'airbnb', 'prettier'],
+  extends: ['plugin:react/recommended', 'airbnb', 'prettier'],
 
-	globals: {
-		Atomics: 'readonly',
-		SharedArrayBuffer: 'readonly',
-	},
+  globals: {
+    Atomics: 'readonly',
+    SharedArrayBuffer: 'readonly',
+  },
 
-	parserOptions: {
-		ecmaFeatures: {
-			jsx: true,
-		},
-		ecmaVersion: 2020,
-		sourceType: 'module',
-	},
+  parserOptions: {
+    ecmaFeatures: {
+      jsx: true,
+    },
+    ecmaVersion: 2020,
+    sourceType: 'module',
+  },
 
-	plugins: ['react', 'prettier', 'react-hooks', 'jsx-a11y'],
+  plugins: ['react', 'prettier', 'react-hooks', 'jsx-a11y'],
 
-	rules: {
-		'linebreak-style': ['error', 'unix'],
-		'no-console': ['error', { allow: ['warn', 'error'] }],
-		curly: ['error', 'multi'],
-		indent: ['error', 2],
-		quotes: ['error', 'single'],
-		semi: ['error', 'always'],
-		'dot-location': ['error', 'property'],
-		eqeqeq: ['error', 'always'],
-		'react/jsx-filename-extension': 0,
-		'react/jsx-indent': [2, 2, { checkAttributes: true, indentLogicalExpressions: true }],
-		'prettier/prettier': ['error', prettierOptions],
-		'import/no-extraneous-dependencies': 0,
-		'import/no-unresolved': [0, { caseSensitive: false }],
-	},
+  rules: {
+    'linebreak-style': ['error', 'unix'],
+    'no-console': ['error', { allow: ['warn', 'error'] }],
+    curly: ['error', 'multi'],
+    indent: ['error', 2],
+    quotes: ['error', 'single'],
+    semi: ['error', 'always'],
+    'dot-location': ['error', 'property'],
+    eqeqeq: ['error', 'always'],
+    'react/jsx-filename-extension': 0,
+    'react/jsx-indent': [2, 2, { checkAttributes: true, indentLogicalExpressions: true }],
+    'prettier/prettier': ['error', prettierOptions],
+    'import/no-extraneous-dependencies': 0,
+    'import/no-unresolved': [0, { caseSensitive: false }],
+  },
 };
