+++ templates/javascript/.gitignore
@@ -26,4 +26,5 @@ yarn-error.log*
 
 yarn.lock
 
+# lint
 report.json
