+++ templates/javascript/.prettierrc
@@ -1,10 +1,10 @@
 {
-	"printWidth": 100,
-	"tabWidth": 2,
-	"useTabs": false,
-	"semi": true,
-	"singleQuote": true,
-	"trailingComma": "all",
-	"arrowParens": "always",
-	"bracketSpacing": true
+  "printWidth": 100,
+  "tabWidth": 2,
+  "useTabs": false,
+  "semi": true,
+  "singleQuote": true,
+  "trailingComma": "all",
+  "arrowParens": "always",
+  "bracketSpacing": true
 }
