+++ templates/javascript/babel.config.js
@@ -1,15 +1,15 @@
 module.exports = {
-	presets: [['@babel/preset-env', { modules: false }], '@babel/preset-react'],
+  presets: [['@babel/preset-env', { modules: false }], '@babel/preset-react'],
 
-	plugins: [
-		'react-hot-loader/babel',
-		'@babel/plugin-proposal-class-properties',
-		'@babel/plugin-syntax-dynamic-import',
-	],
+  plugins: [
+    'react-hot-loader/babel',
+    '@babel/plugin-proposal-class-properties',
+    '@babel/plugin-syntax-dynamic-import',
+  ],
 
-	env: {
-		test: {
-			plugins: ['@babel/plugin-transform-modules-commonjs', 'dynamic-import-node'],
-		},
-	},
+  env: {
+    test: {
+      plugins: ['@babel/plugin-transform-modules-commonjs', 'dynamic-import-node'],
+    },
+  },
 };
