+++ templates/javascript/src/bootstrap.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import App from './app';
+import App from 'pages/App';
 import 'styles/main.css';
 
 const root = document.getElementById('root');
