+++ templates/javascript/src/pages/App.js
