+++ templates/javascript/src/styles/app.scss
@@ -1,30 +1,30 @@
 .wrapper {
-	width: 100%;
-	height: 100%;
-	display: flex;
-	flex-direction: column;
-	justify-content: center;
-	align-items: center;
-	background: #0d1b2b;
-	box-sizing: border-box;
+  width: 100%;
+  height: 100%;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  background: #0d1b2b;
+  box-sizing: border-box;
 }
 
 .title {
-	font-size: 60px;
-	color: #457;
+  font-size: 60px;
+  color: #457;
 }
 
 .status {
-	width: 50%;
-	display: flex;
-	justify-content: center;
-	align-items: center;
-	border: 10px solid #457;
-	border-radius: 15px;
-	padding: 20px;
+  width: 50%;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  border: 10px solid #457;
+  border-radius: 15px;
+  padding: 20px;
 }
 
 .status > span {
-	font-size: 50px;
-	color: #457b9d;
+  font-size: 50px;
+  color: #457b9d;
 }
