+++ templates/javascript/src/styles/main.css
@@ -1,9 +1,9 @@
 html,
 body,
 #root {
-	width: 100%;
-	height: 100%;
-	margin: 0;
-	padding: 0;
-	box-sizing: border-box;
+  width: 100%;
+  height: 100%;
+  margin: 0;
+  padding: 0;
+  box-sizing: border-box;
 }
